init gang

+++ utils.c
@@ -6,15 +6,15 @@
 /*   By: jcardina <succosopompelmo>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/26 17:10:26 by jcardina          #+#    #+#             */
-/*   Updated: 2023/09/26 17:18:37 by jcardina         ###   ########.fr       */
+/*   Updated: 2023/09/28 12:33:31 by jcardina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int		i;
+	int		sign;
+	long	result;
 
 	i = 0;
 	while ((str[i] >= 9 && str[i] <= 13) || (str[i] == 32))
